Moving Netcdffilebuffer to a new class file
